move utility classes to utils package
